Move the satellite obs to the obs file
