+++ mlpp_features/__init__.py
@@ -4,4 +4,3 @@ from mlpp_features.nwp import *
 from mlpp_features.obs import *
 from mlpp_features.terrain import *
 from mlpp_features.time import *
-from mlpp_features.sat import *
